requirements.txt
pandas
numpy
Pillow
matplotlib
imagesize
sklearn
torch
transformers